Fixing codeclimate suggestions
